+++ scimap/helpers/__init__.py
@@ -1,1 +1,1 @@
-from ._rename import renamefrom ._classify import classify
+from ._rename import renamefrom ._classify import classifyfrom ._scimap_to_csv import scimap_to_csv